move datetime to pkg
