+++ GitlabCreateAccessTokenOptions.go
@@ -4,8 +4,8 @@ import (
 	"time"
 
 	"github.com/asciich/asciichgolangpublic/datatypes/slicesutils"
-	"github.com/asciich/asciichgolangpublic/datetime"
 	"github.com/asciich/asciichgolangpublic/logging"
+	"github.com/asciich/asciichgolangpublic/pkg/datetime"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 )
