+++ pkg/tlsutils/x509utils/X509CreateCertificateOptions.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	"github.com/asciich/asciichgolangpublic/datatypes/bigintutils"
-	"github.com/asciich/asciichgolangpublic/datetime"
 	"github.com/asciich/asciichgolangpublic/logging"
+	"github.com/asciich/asciichgolangpublic/pkg/datetime"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 )
